backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.7
requests==2.31.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
networkx==3.2.1
chromadb==0.4.22
numpy==1.26.4
google-generativeai==0.5.0
PyMuPDF==1.23.8


